Feat: added kuwahara Picker to sample App Feat: KuwaharaTypes now Support case Iterable and has a title for each type
